+++ frontend/src/core/features/SpotlightSearch/SpotlightSearch.tsx
@@ -37,6 +37,7 @@ import useOnClickOutside from "use-onclickoutside";
 import { max } from "lodash";
 import InputSearch from "./InputSearch";
 import Spinner from "core/components/Spinner";
+import { GetServerSidePropsContext } from "next";
 
 type Workspace = GetWorkspacesQuery["workspaces"]["items"][0];
 
@@ -57,11 +58,9 @@ const getTabLabel = (label: string, totalItems?: number): string => {
 const pageSize = 15;
 
 const SpotlightSearch = ({
-  isMac,
   isSidebarOpen,
   organizationId,
 }: {
-  isMac: boolean;
   isSidebarOpen: boolean;
   organizationId?: string;
 }) => {
@@ -335,7 +334,6 @@ const SpotlightSearch = ({
     return (
       <InputSearch
         isSidebarOpen={isSidebarOpen}
-        isMac={isMac}
         onClick={() => setIsOpen((prev) => !prev)}
       />
     );
@@ -460,7 +458,6 @@ const SpotlightSearch = ({
       </div>
       <InputSearch
         isSidebarOpen={isSidebarOpen}
-        isMac={isMac}
         onClick={() => setIsOpen((prev) => !prev)}
       />
     </>
@@ -535,6 +532,10 @@ SpotlightSearch.fragments = {
       ${WorkspaceDisplay.fragments.workspace}
     }
   `,
+};
+
+SpotlightSearch.prefetch = async (ctx: GetServerSidePropsContext) => {
+  await InputSearch.prefetch(ctx);
 };
 
 export default SpotlightSearch;
